pyproject.toml
[tool.poetry]
name = "qa-guru-lesson-19"
version = "0.1.0"
description = ""
authors = ["Irina.Voron <yarutich.irina@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
selene = "2.0.0rc9"
appium-python-client = "^4.0.0"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"
allure-pytest = "^2.13.5"
pydantic-settings = "^2.2.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
